Refactor integration test. Marked two unittest as blocked. See: https://kb-dk.atlassian.net/browse/DRA-1773
