+++ src/test/java/dk/kb/discover/SolrServiceTest.java
@@ -39,44 +39,7 @@ import static org.mockito.Mockito.mockStatic;
 class SolrServiceTest {
 
     private static final Logger log = LoggerFactory.getLogger(SolrServiceTest.class);
-    
-    private static DsDiscoverClient remote = null;
-    private static String dsDiscoverDevel=null;  
-    
-    @BeforeAll
-    static void setUp() throws Exception{
-        try {
-            ServiceConfig.initialize("ds-discover-integration-test.yaml"); 
-            dsDiscoverDevel= ServiceConfig.getConfig().getString("discover.url");
-            remote = new DsDiscoverClient(dsDiscoverDevel);
-        } catch (IOException e) { 
-            e.printStackTrace();
-            log.error("Integration yaml 'ds-discover-integration-test.yaml' file most be present. Call 'kb init'"); 
-            fail();
-        }
         
-        try {            
-            String keyCloakRealmUrl= ServiceConfig.getConfig().getString("integration.devel.keycloak.realmUrl");            
-            String clientId=ServiceConfig.getConfig().getString("integration.devel.keycloak.clientId");
-            String clientSecret=ServiceConfig.getConfig().getString("integration.devel.keycloak.clientSecret");                
-            String token=KeycloakUtil.getKeycloakAccessToken(keyCloakRealmUrl, clientId, clientSecret);           
-            log.info("Retrieved keycloak access token:"+token);            
-            
-            //Mock that we have a JaxRS session with an Oauth token as seen from within a service call.
-            MessageImpl message = new MessageImpl();                            
-            message.put(OAuthConstants.ACCESS_TOKEN_STRING,token);            
-            MockedStatic<JAXRSUtils> mocked = mockStatic(JAXRSUtils.class);           
-            mocked.when(JAXRSUtils::getCurrentMessage).thenReturn(message);
-                                                                         
-        }
-        catch(Exception e) {
-            log.warn("Could not retrieve keycloak access token. Service will be called without Bearer access token");            
-            e.printStackTrace();
-        }                        
-    }
-    
-    // Integration test: Requires a local solr at port 10007 from the Digitale Samlinger project with
-    // a ds-collection
     //@Test
     void baseSearch() {
         SolrService solr = new SolrService("test", "http://localhost:10007", "solr", "ds");
@@ -174,51 +137,6 @@ class SolrServiceTest {
                 () -> SolrService.removePrefixedFilters(response, prefix,"xml"));
     }
 
-    @Test
-    @Tag("integration")
-    void suggestTestNotAvailable() throws IOException {
-        // Integration test towards devel env. Remember to update aegis before running this.
-        ServiceConfig.initialize("src/test/resources/ds-discover-integration-test.yaml");
-        String suggestDictionary = "radiotv_title_suggest";
-        // no suggestions should be available for this query.
-        String suggestQuery = "Palle Lauring";
-        int suggestCount = 5;
-        String wt = "json";
-        SolrService solr = SolrManager.getSolrService("ds");
-
-        String filteredResponse = solr.suggest(suggestDictionary, suggestQuery, suggestCount, wt);
-
-        assertTrue(filteredResponse.contains("\"suggest\" : {\n" +
-                "    \"radiotv_title_suggest\" : {\n" +
-                "      \"Palle Lauring\" : {\n" +
-                "        \"numFound\" : 0,\n" +
-                "        \"suggestions\" : [ ]\n" +
-                "      }\n" +
-                "    }\n" +
-                "  }"));
-    }
-
-    @Test
-    @Tag("integration")
-    void suggestTest() throws IOException {
-        // Integration test towards devel env. Remember to update aegis before running this.
-        ServiceConfig.initialize("src/test/resources/ds-discover-integration-test.yaml");
-        String suggestDictionary = "radiotv_title_suggest";
-        // no suggestions should be available for this query.
-        String suggestQuery = "tes";
-        int suggestCount = 5;
-        String wt = "json";
-        SolrService solr = SolrManager.getSolrService("ds");
-
-        String filteredResponse = solr.suggest(suggestDictionary, suggestQuery, suggestCount, wt);
-
-        System.out.println(filteredResponse);
-
-        assertTrue(filteredResponse.contains("\"suggest\" : {\n" +
-                "    \"radiotv_title_suggest\" : {\n" +
-                "      \"tes\" : {\n" +
-                "        \"numFound\" : 8,\n"));  //This number will change depending on corpus
-    }
-
+  
 
 }
