+++ src/test/java/dk/kb/discover/util/integration/DsDiscoverClientTest.java
@@ -1,0 +1,26 @@
+package dk.kb.discover.util.integration;
+
+import java.io.IOException;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Tag;
+import org.junit.jupiter.api.Test;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+
+/**
+ * Simple verification of client code generation.
+ */
+
+@Tag("integration")
+public class DsDiscoverClientTest extends IntegrationTest{
+    private static final Logger log = LoggerFactory.getLogger(DsDiscoverClientTest.class);
+   
+    @Test
+    public void test() throws IOException {
+    //Must be one unit test to test the setup method is working 
+    }
+         
+         
+    
+}
